requirements.txt
streamlit>=1.30
pytest>=7.0
huggingface_hub>=0.24
pydantic>=2.0
python-dotenv>=1.0
